Support modern v4 PuppetDB query parameters
With PuppetDB query v4 all of the parameters were adjusted to
expect underscores. Additionally, there were other parameters
which expected to receive structured data like "counts-filter".
This generalizes the support so that any array or hash values
are serialized as JSON, which should expand compatibility to
all the currently supported parameters.